Force removal of (IRC) suffix in nick completion
In many places with matrix<->irc bridge usage,
it appends (IRC) suffix to user name,
which is just terrible to read on irc side,
where majority knows they are on irc,
and bots do not like such garbage in the messages.

So to make usage of Quaternion more comfortable
in such networks, make sure tab completion
never puts (IRC) suffix to the text message sent to the room.